Adjust ROM fingerprint to change regularly
The purpose of the ROM fingerprint is so that we can use it to trigger a
wipe of the package cache. If the ROM fingerprint only changes once per day,
this doesn't work for us (developers) who compile sometimes up to 10-20 in
one single day. Lets add the time to the ROM fingerprint so that it triggers
the wipe on dirty flashes.

Signed-off-by: calebcabob <calphonic@gmail.com>